Initial commit - adapted from coordinate-window
A small application that creates a window and displays a zoomed in
portion of the screen around the cursor to it.

Presently the zoom factor is fixed at 3x.

Polling is used to determine when the cursor position has changed.
The update interval is somewhat smart, such that it will reduce the
interval if there have not been any changes to it within 1 second.
The fast interval is 32ms (or 1/30th of a second).